fixed warehouse utils to warehouse func
